+++ h2/src/test/org/h2/test/db/TestListener.java
@@ -74,7 +74,7 @@ public class TestListener extends TestDb implements DatabaseEventListener {
     }
 
     @Override
-    public void setProgress(int state, String name, int current, int max) {
+    public void setProgress(int state, String name, long current, long max) {
         long time = System.nanoTime();
         if (state == lastState && time < last + TimeUnit.SECONDS.toNanos(1)) {
             return;
